[iterator.intersection.local] fixed major bug
The local intersection ids that we are to act on were provided as a set,
thus ordered numerically. But since the intersection iterator traverses the
intersections in a different order we would assign last_ to the highest
intersection number and not to the number of the last intersection. Making
this a vector preserves the order...